# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /advogados/contagem:
    get:
      operationId: contarAdvogados
      tags: [cadastro]
      summary: Count active, visible lawyers by area and location
      description: >-
        Returns counts of lawyers with an active subscription and a complete
        profile, optionally filtered by area, city and state. Used to show real
        competition numbers during the registration funnel.
      parameters:
        - name: area
          in: query
          required: false
          schema:
            type: string
        - name: cidade
          in: query
          required: false
          schema:
            type: string
        - name: uf
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Competition counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConcorrenciaResult"
...
  /perfil:
    get:
      operationId: getPerfil
      tags: [lawyer]
      summary: Get the authenticated lawyer's profile
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LawyerProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
    put:
      operationId: updatePerfil
      tags: [lawyer]
      summary: Create or update the authenticated lawyer's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LawyerProfile"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Cidade:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        uf:
          type: string
          minLength: 2
          maxLength: 2
      required:
        - nome
        - uf
      additionalProperties: false
    LawyerProfile:
      type: object
      properties:
        nome:
          type: string
        oab:
          type: string
        photo:
          type: string
          nullable: true
        about:
          type: string
        areas:
          type: array
          items:
            type: string
        cidades:
          type: array
          items:
            $ref: "#/components/schemas/Cidade"
        atendeOnline:
          type: boolean
        whatsapp:
          type: string
        instagram:
          type: string
        linkedin:
          type: string
        website:
          type: string
        outro:
          type: string
        complete:
          type: boolean
        subscriptionStatus:
          type: string
          enum: [pendente, ativa, atrasada, inativa]
          nullable: true
        visivel:
          type: boolean
        oabVerificada:
          type: boolean
        oabSituacao:
          type: string
          nullable: true
        oabNomeConfirmado:
          type: string
          nullable: true
        oabVerificadaEm:
          type: string
          nullable: true
        oabVerificacaoPendente:
          type: boolean
        createdAt:
          type: string
          nullable: true
      required:
        - nome
        - oab
        - about
        - areas
        - cidades
        - atendeOnline
        - whatsapp
        - instagram
        - linkedin
        - website
        - outro
        - complete
        - visivel
        - oabVerificada
        - oabVerificacaoPendente
    UpdateProfileInput:
      type: object
      properties:
        nome:
          type: string
        oab:
          type: string
        photo:
          type: string
          nullable: true
        about:
          type: string
        areas:
          type: array
          items:
            type: string
        cidades:
          type: array
          items:
            $ref: "#/components/schemas/Cidade"
        atendeOnline:
          type: boolean
        whatsapp:
          type: string
        instagram:
          type: string
        linkedin:
          type: string
        website:
          type: string
        outro:
          type: string
        # Token de verificação da OAB assinado pelo servidor (emitido por
        # POST /verificar-oab quando valido=true). O status "verificado" é
        # server-managed: o cliente NÃO grava oabVerificada diretamente, apenas
        # apresenta este token, que o servidor valida antes de marcar verificado.
        oabToken:
          type: string
          nullable: true
        # Sinaliza verificação pendente (revisão manual) quando o serviço da OAB
        # falhou. É o estado de MENOR privilégio, então pode vir do cliente.
        oabVerificacaoPendente:
          type: boolean
      required:
        - nome
        - oab
        - about
        - areas
        - cidades
        - atendeOnline
        - whatsapp
        - instagram
        - linkedin
        - website
        - outro
      additionalProperties: false
...
    ConcorrenciaResult:
      type: object
      properties:
        total:
          type: integer
        naArea:
          type: integer
        naCidade:
          type: integer
        naAreaECidade:
          type: integer
      required:
        - total
        - naArea
        - naCidade
        - naAreaECidade
...